# Test/rust scanner