--- a/Dockerfile_build
+++ b/Dockerfile_build
@@ -15,7 +15,7 @@ RUN apt-get update && \
     autoconf automake autotools-dev libtool xutils-dev valgrind && \
     rm -rf /var/lib/apt/lists/*
 
-# Download, verify, and install ragel version 6.
+# Download, verify, and install ragel version 6. 
 ENV RAGEL_VERSION=6.10
 RUN curl https://www.colm.net/files/thurston.asc | gpg --import - && \
     curl https://www.colm.net/files/ragel/ragel-${RAGEL_VERSION}.tar.gz -O && \