--- a/libflux/core/Cargo.toml
+++ b/libflux/core/Cargo.toml
@@ -30,13 +30,7 @@ tabwriter = "1.2.1"
 [dev-dependencies]
 colored = "1.8"
 pretty_assertions = "0.6.1"
-criterion = "0.3.3"
 
 [build-dependencies]
 cc = "1.0"
 bindgen = "0.49.0"
-
-[[bench]]
-name = "bench_scanner"
-path = "scanner/benchmarks.rs"
-harness = false