--- a/libflux/core/src/scanner/mod.rs
+++ b/libflux/core/src/scanner/mod.rs
@@ -44,6 +44,19 @@ pub struct Token {
     pub comments: Option<Box<Token>>,
 }
 
+// To use the `{}` marker, the trait `fmt::Display` must be implemented
+// manually for the type.
+impl fmt::Display for Token {
+    // This trait requires `fmt` with this exact signature.
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        // Write strictly the first element into the supplied output
+        // stream: `f`. Returns `fmt::Result` which indicates whether the
+        // operation succeeded or failed. Note that `write!` uses syntax which
+        // is very similar to `println!`.
+        write!(f, "{}", self.lit)
+    }
+}
+
 impl Scanner {
     // New creates a scanner with the provided input.
     pub fn new(data: CString) -> Scanner {