--- a/libflux/core/src/scanner/mod.rs
+++ b/libflux/core/src/scanner/mod.rs
@@ -3,6 +3,7 @@ include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 
 pub type CChar = u8;
 
+use crate::fmt;
 use std::collections::HashMap;
 use std::ffi::CString;
 use std::str;