Closes #3267. 

This issue measures performance changes after porting the scanner to rust. We were just looking to not regress performance, but there are some modest gains as seen below. 

|test name|old|new|performance change|
|--|--|-----|---|
|scan_short_text|60.383 ns|44.630 ns| -26.5% | 
|scan_long_text|45.097 ns|43.696 ns| -3.1% |
|replaceAll|456.28 us|383.14 us| -16.0% |
|title_test|442.05 us|363.45 us| -17.8% |
|trim_test|390.74 us|373.56 us| -4.4% |
|toUpper|377.92 us|374.77 us| -0.8% |
|substring|447.97 us|421.40 us| -5.9% |
|toLower|383.53 us|378.75 us| -1.2% |
|replace_test|454.77 us|399.19 us | -12.2% |
|length_test|505.50 us| 382.15 us| -24.4% |
|subset_test|445.94 us| 389.79 us | -12.6% |